Added support in the Cameras_Tests file to run on my computer. All tests from LG_Cameras_Tests are working ok

+++ LAIG-Proj1/LG_Node.cpp
@@ -77,5 +77,5 @@ string LG_Node::getIdentifier(){
 
 int LG_Node::getNrChildren()
 {
-	return this->childsIDs.size();
+	return (int) this->childsIDs.size();
 }
